+++ compiler/llvmir/compile_test.go
@@ -1,29 +1,21 @@
 package llvmir
 
 import (
-	"fmt"
+	"io"
 	"io/ioutil"
 	"os"
 	"os/exec"
 	"strings"
 	"testing"
-
-	"github.com/llir/llvm/ir"
 )
 
-func parseTestFile(t *testing.T, testcase string) *ir.Module {
+func loadTestFile(t *testing.T, testcase string) io.ReadCloser {
 	t.Helper()
 	f, err := os.Open("../../testsuite/" + testcase + ".l")
 	if err != nil {
 		t.Fatal(err)
 	}
-	defer f.Close()
-	m, err := Generate(f)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	return m
+	return f
 }
 
 func runWithArgs(t *testing.T, name, dir string, args []string, estdout, estderr string) {
@@ -63,226 +55,18 @@ func runWithArgs(t *testing.T, name, dir string, args []string, estdout, estderr
 	}
 }
 
-// Compile a program and return the dir that was used as a temporary dir.
-// It's the caller's responsibility to clean up dir.
-func compile(t *testing.T, name string) string {
-	m := parseTestFile(t, name)
-	dir, err := ioutil.TempDir("", "langtest"+name+"_")
-	if err != nil {
-		t.Fatal(err)
-	}
-	f, err := os.Create(dir + "/usercode.ll")
-	if err != nil {
-		t.Fatal(err)
-	}
-	runtime, err := os.Create(dir + "/runtime.ll")
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	fmt.Fprintf(f, "%v", m)
-
-	// FIXME: The builtins that have been re-writing in stdlib should be used instead of these.
-	fmt.Fprintf(runtime, "%v", startSymbol(m)+`
-
-	define void @Exit(i64 %code) {
-		call void asm sideeffect "movq $$`+SYS_EXIT+`, %rax\0Asyscall\0A", "{di}"(i64 %code)
-		unreachable
-	}
-	define void @PrintString({i8*, i64} %str) {
-		tail call void @Write(i64 1, {i8*, i64} %str)
-		ret void
-	}
-
-	define void @Write(i64 %fd, {i8*, i64} %buf) {
-		%base = extractvalue {i8*, i64} %buf, 0
-		%size = extractvalue {i8*, i64} %buf, 1
-		tail call void asm sideeffect "movq $$`+SYS_WRITE+`, %rax\0Asyscall\0A", "{di},{dx},{si},~{dirflag},~{fpsr},~{flags}"(i64 %fd, i64 %size, i8* %base)
-		ret void
-	}
-
-	define i64 @Read(i64 %fd, {i8*, i64} %buf) {
-		%base = extractvalue {i8*, i64} %buf, 0
-		%size = extractvalue {i8*, i64} %buf, 1
-		%1 = call i32 asm sideeffect "movq $$`+SYS_READ+`, %rax\0Asyscall\0A", "=A,{di},{dx},{si},~{dirflag},~{fpsr},~{flags}"(i64 %fd, i64 %size, i8* %base)
-		%2 = sext i32 %1 to i64
-		ret i64 %2
-	}
-
-
-	define i64 @Create({i8*, i64} %buf) {
-		%base = extractvalue {i8*, i64} %buf, 0
-		%size = extractvalue {i8*, i64} %buf, 1
-		%1 = call i32 asm sideeffect "`+strings.Replace(`movq %rcx, %rax
-		movq %rsp,%rdi
-		subq %rcx, %rdi
-		cld
-		rep movsb
-		subq %rax, %rdi
-movq $$438, %rdx
-movq $$`+CREATE_CONST+`, %rsi
-movq $$`+SYS_OPEN+`, %rax
-syscall`, "\n", `\0A`, -1)+`", "=A,{si},{cx}"(i8* %base, i64 %size) 
-		%2 = sext i32 %1 to i64
-		ret i64 %2
-	}
-
-	define i64 @Open ({i8*, i64} %buf) {
-		%base = extractvalue {i8*, i64} %buf, 0
-		%size = extractvalue {i8*, i64} %buf, 1
-		%1 = call i32 asm sideeffect "`+strings.Replace(`movq %rcx, %rax
-		movq %rsp,%rdi
-		subq %rcx, %rdi
-		subq $$16, %rdi`+ // FIXME: This should be aligned, this is just so that movsb doesn't
-		// overwrite the return address on the stack.
-		`
-		cld
-		rep movsb
-		subq %rax, %rdi
-		movb $$0, (%rdi, %rax)
-movq $$0, %rsi
-movq $$0, %rdx
-movq $$`+SYS_OPEN+`, %rax
-syscall`, "\n", `\0A`, -1)+`", "=A,{si},{cx},~{dirflag},~{fpsr},~{flags}"(i8* %base, i64 %size)
-		%2 = sext i32 %1 to i64
-		ret i64 %2
-	}
-
-
-	define void @Close(i64 %fd) {
-		tail call void asm sideeffect "movq $$`+SYS_CLOSE+`, %rax\0Asyscall\0A", "{di},~{dirflag},~{fpsr},~{flags}"(i64 %fd)
-		ret void
-	}
-	
-	define void @PrintInt(i64 %n) {
-		%1 = alloca [20 x i8]
-		%reversed = alloca [20 x i8]
-		%rem = alloca i64
-		%len = alloca i8
-		%i = alloca i8
-		store i64 %n, i64* %rem
-		store i8 0, i8* %len
-		store i8 0, i8* %i
-
-		%zero = icmp eq i64 0, %n
-		br i1 %zero, label %printzero, label %printnonzero
-		printzero:
-		%rfirst = getelementptr [20 x i8], [20 x i8]* %reversed, i8 0, i8 0
-		store i8 48, i8* %rfirst
-		store i8 1, i8* %len
-		br label %print
-		printnonzero:
-		%isneg = icmp slt i64 %n, 0
-		br i1 %isneg, label %neg, label %pos
-		neg:
-		%abs = mul i64 -1, %n
-		store i64 %abs, i64* %rem
-		br label %adddigit
-		pos:
-		store i64 %n, i64* %rem
-		br label %adddigit
-		adddigit:
-		%idx = load i8, i8* %len
-		%el = getelementptr [20 x i8], [20 x i8]* %1, i8 0, i8 %idx
-
-		%x = load i64, i64* %rem
-		%digit = srem i64 %x, 10
-		%trunc = trunc i64 %digit to i8
-		%dchar = add i8 48, %trunc
-		store i8 %dchar, i8* %el
-
-		%newlen = add i8 1, %idx
-		store i8 %newlen, i8* %len
-
-		%newval = sdiv i64 %x, 10
-		store i64 %newval, i64* %rem
-
-		%done = icmp eq i64 0, %newval
-		br i1 %done, label %checksign, label %adddigit
-		checksign:
-		br i1 %isneg, label %addsign, label %reverse
-		addsign:
-		%newlen2 = add i8 1, %idx
-		store i8 %newlen2, i8* %len
-		%sel = getelementptr [20 x i8], [20 x i8]* %1, i8 0, i8 %newlen2
-		store i8 45, i8* %sel
-		%newlen3 = add i8 1, %newlen2
-		store i8 %newlen3, i8* %len
-		br label %reverse
-		reverse:
-		%finallen = load i8, i8* %len
-		%iv = load i8, i8* %i
-
-		%ridx = getelementptr [20 x i8], [20 x i8]* %reversed, i8 0, i8 %iv
-		%2 = sub i8 %finallen, %iv
-		%negidx = sub i8 %2, 1
-
-		%rdigit = getelementptr [20 x i8], [20 x i8]* %1, i8 0, i8 %negidx
-		%dgt = load i8, i8* %rdigit
-
-		store i8 %dgt, i8* %ridx
-
-		%addi = add i8 1, %iv
-		store i8 %addi, i8* %i
-
-		%donerev = icmp eq i8 %addi, %finallen
-		br i1 %donerev, label %print, label %reverse
-		print:
-		%lenval = load i8, i8* %len
-		%zexlen = zext i8 %lenval to i64
-		%first = getelementptr [20 x i8], [20 x i8]* %reversed, i8 0, i8 0
-		%ptr = insertvalue {i8*, i64} {i8* undef, i64 undef}, i8* %first, 0
-		%ptrwlen = insertvalue {i8*, i64} %ptr, i64 %zexlen, 1
-		call void @Write(i64 1, {i8*, i64} %ptrwlen)
-		ret void
-	}
-
-
-	`)
-
-	cmd := exec.Command("llc", "-filetype=obj", dir+"/runtime.ll")
-	if err := cmd.Start(); err != nil {
-		t.Errorf("%v: llc: %v", name, err)
-		return dir
-	}
-	if err := cmd.Wait(); err != nil {
-		t.Errorf("%v: llc: %v", name, err)
-		return dir
-	}
-	cmd = exec.Command("llc", "-filetype=obj", dir+"/usercode.ll")
-	if err := cmd.Start(); err != nil {
-		t.Errorf("%v: llc: %v", name, err)
-		return dir
-	}
-	if err := cmd.Wait(); err != nil {
-		t.Errorf("%v: llc: %v", name, err)
-		return dir
-	}
-	cmd = exec.Command("llc", "-filetype=obj", dir+"/usercode.ll")
-	if err := cmd.Start(); err != nil {
-		t.Errorf("%v: llc: %v", name, err)
-		return dir
-	}
-	if err := cmd.Wait(); err != nil {
-		t.Errorf("%v: llc: %v", name, err)
-		return dir
-	}
-	cmd = exec.Command("ld", "-o", dir+"/main", dir+"/usercode.o", dir+"/runtime.o")
-	if err := cmd.Start(); err != nil {
-		t.Errorf("%v: ld: %v", name, err)
-		return dir
-	}
-	if err := cmd.Wait(); err != nil {
-		t.Errorf("%v: ld: %v", name, err)
-		return dir
-	}
-	return dir
-}
-
 func compileAndRun(t *testing.T, name string, estdout, estderr string) {
 	t.Helper()
-	dir := compile(t, name)
-	defer os.RemoveAll(dir)
+	r := loadTestFile(t, name)
+	defer r.Close()
+
+	dir, err := Compile(r, false)
+	if dir != "" {
+		defer os.RemoveAll(dir)
+	}
+	if err != nil {
+		t.Fatal(err)
+	}
 	runWithArgs(t, name, dir, nil, estdout, estderr)
 }
 
@@ -397,8 +181,14 @@ func TestCreateSyscall(t *testing.T) {
 		os.Chdir(pwd)
 	}(pwd)
 
-	dir := compile(t, "createsyscall")
-	defer os.RemoveAll(dir)
+	r := loadTestFile(t, "createsyscall")
+	dir, err := Compile(r, false)
+	if dir != "" {
+		defer os.RemoveAll(dir)
+	}
+	if err != nil {
+		t.Fatal(err)
+	}
 
 	// Make sure foo gets created in dir, so that the defer cleans it up.
 	if err := os.Chdir(dir); err != nil {
@@ -428,7 +218,14 @@ func TestReadSyscall(t *testing.T) {
 		os.Chdir(pwd)
 	}(pwd)
 
-	dir := compile(t, "readsyscall")
+	r := loadTestFile(t, "readsyscall")
+	dir, err := Compile(r, false)
+	if dir != "" {
+		defer os.RemoveAll(dir)
+	}
+	if err != nil {
+		t.Fatal(err)
+	}
 	defer os.RemoveAll(dir)
 
 	// Make sure foo gets created in dir, so that the defer cleans it up.
@@ -458,7 +255,14 @@ func TestArgc(t *testing.T) {
 		os.Chdir(pwd)
 	}(pwd)
 
-	dir := compile(t, "argc")
+	r := loadTestFile(t, "argc")
+	dir, err := Compile(r, false)
+	if dir != "" {
+		defer os.RemoveAll(dir)
+	}
+	if err != nil {
+		t.Fatal(err)
+	}
 	defer os.RemoveAll(dir)
 
 	runWithArgs(t, "argc", dir, nil, "1", "")
@@ -477,7 +281,14 @@ func TestArgLens(t *testing.T) {
 		os.Chdir(pwd)
 	}(pwd)
 
-	dir := compile(t, "arglens")
+	r := loadTestFile(t, "arglens")
+	dir, err := Compile(r, false)
+	if dir != "" {
+		defer os.RemoveAll(dir)
+	}
+	if err != nil {
+		t.Fatal(err)
+	}
 	defer os.RemoveAll(dir)
 
 	runWithArgs(t, "arglens", dir, nil, "\n", "")
@@ -495,8 +306,14 @@ func TestEchoProgram(t *testing.T) {
 		os.Chdir(pwd)
 	}(pwd)
 
-	dir := compile(t, "echo")
-	defer os.RemoveAll(dir)
+	r := loadTestFile(t, "echo")
+	dir, err := Compile(r, false)
+	if dir != "" {
+		defer os.RemoveAll(dir)
+	}
+	if err != nil {
+		t.Fatal(err)
+	}
 
 	runWithArgs(t, "echo", dir, []string{"foo", "bar"}, "foo bar\n", "")
 	runWithArgs(t, "echo", dir, []string{"other", "params"}, "other params\n", "")
@@ -521,8 +338,14 @@ func TestCatPrograms(t *testing.T) {
 			t.Fatal(err)
 		}
 
-		dir := compile(t, c)
-		defer os.RemoveAll(dir)
+		r := loadTestFile(t, c)
+		dir, err := Compile(r, false)
+		if dir != "" {
+			defer os.RemoveAll(dir)
+		}
+		if err != nil {
+			t.Fatal(err)
+		}
 		// Make sure foo gets created in dir, so that the defer cleans it up, now
 		// that we've parsed the program.
 		if err := os.Chdir(dir); err != nil {
